Implemented Some of BL

+++ AcademyProject/AcademyProjectSL/Services/AuthorService.cs
@@ -11,54 +11,61 @@ namespace AcademyProjectSL.Services
     {
         private readonly IMapper _mapper;
         private readonly IAuthorInMemoryRepo _authorInMemoryRepo;
+        private readonly IBookService _bookService;
 
-        public AuthorService(IAuthorInMemoryRepo authorInMemoryRepo, IMapper mapper)
+        public AuthorService(IAuthorInMemoryRepo authorInMemoryRepo, IMapper mapper, IBookService bookService)
         {
             _mapper = mapper;
             _authorInMemoryRepo = authorInMemoryRepo;
+            _bookService = bookService;
         }
 
-        public IEnumerable<Author> GetAuthors => _authorInMemoryRepo.GetAuthors;
+        public async Task<IEnumerable<Author>> GetAuthors() => await _authorInMemoryRepo.GetAuthors();
 
-        public Author? GetById(int id)
+        public async Task<Author?> GetById(int id)
         {
-            return _authorInMemoryRepo.GetById(id);
+            if (id <= 0) return null;
+            return await _authorInMemoryRepo.GetById(id);
         }
 
-        public AddAuthorResponse AddAuthor(AddAuthorRequest authorRequest)
+        public async Task<AddAuthorResponse> AddAuthor(AddAuthorRequest authorRequest)
         {
             if (authorRequest == null) return new AddAuthorResponse { HttpStatusCode = System.Net.HttpStatusCode.BadRequest, Message = "Request is null" };
 
-            var result = _authorInMemoryRepo.GetAuthorByName(authorRequest.Name);
+            var result = await _authorInMemoryRepo.GetAuthorByName(authorRequest.Name);
 
             if (result != null) return new AddAuthorResponse() { HttpStatusCode = System.Net.HttpStatusCode.BadRequest, Message = "The Author already exists" };
 
             var author = _mapper.Map<Author>(authorRequest);
-            var result2 = _authorInMemoryRepo.AddAuthor(author);
+            var result2 = await _authorInMemoryRepo.AddAuthor(author);
             return new AddAuthorResponse() { HttpStatusCode = System.Net.HttpStatusCode.OK, Author = result2 };
         }
 
-        public UpdateAuthoreResponse UpdateAuthor(UpdateAuthorRequest authorRequest)
+        public async Task<UpdateAuthoreResponse> UpdateAuthor(UpdateAuthorRequest authorRequest)
         {
             if (authorRequest == null) return new UpdateAuthoreResponse { HttpStatusCode = System.Net.HttpStatusCode.BadRequest, Message = "Request is null" };
 
-            var result = _authorInMemoryRepo.GetById(authorRequest.Id);
+            var result = await _authorInMemoryRepo.GetById(authorRequest.Id);
 
             if (result == null) return new UpdateAuthoreResponse() { HttpStatusCode = System.Net.HttpStatusCode.BadRequest, Message = $"Cant find this Author: {authorRequest.Name}" };
 
             var result2 = _mapper.Map<Author>(authorRequest);
-            _authorInMemoryRepo.UpdateAuthor(result2);
+            await _authorInMemoryRepo.UpdateAuthor(result2);
             return new UpdateAuthoreResponse { HttpStatusCode = System.Net.HttpStatusCode.OK, Author = result2 };
         }
 
-        public Author? DeleteAuthor(int authorId)
+        public async Task<Author?> DeleteAuthor(int authorId)
         {
-            return _authorInMemoryRepo.DeleteAuthor(authorId);
+            if (authorId <= 0) return null;
+            var author = await GetById(authorId);
+            if (author == null) return null;
+            if (await _bookService.GetBookByAuthorId(author.Id) != null) return null;
+            return await _authorInMemoryRepo.DeleteAuthor(authorId);
         }
 
-        public Author? GetAuthorByName(string name)
+        public async Task<Author?> GetAuthorByName(string name)
         {
-            return _authorInMemoryRepo.GetAuthorByName(name);
+            return await _authorInMemoryRepo.GetAuthorByName(name);
         }
     }
 }
